+++ sources/packages/backend/apps/api/src/route-controllers/index.ts
@@ -53,3 +53,4 @@ export * from "./student-account-applications/student-account-application.studen
 export * from "./education-program-offering/education-program-offering.institutions.controller";
 export * from "./education-program-offering/education-program-offering.controller.service";
 export * from "./education-program-offering/education-program-offering.students.controller";
+export * from "./user/user.controller.service";
