+++ sources/packages/backend/apps/api/src/route-controllers/user/user.institutions.controller.ts
@@ -12,16 +12,12 @@ import { IUserToken } from "../../auth/userToken.interface";
 import { BCeIDDetailsAPIOutDTO } from "./models/bceid-account.dto";
 import { InstitutionUserAPIOutDTO } from "./models/institution-user.dto";
 import { InstitutionUserPersistAPIInDTO } from "./models/institution-user-persist.dto";
-import { SearchAccountOptions } from "../../services/bceid/search-bceid.model";
 import { BCeIDAccountsAPIOutDTO } from "./models/bceid-accounts.dto";
 import { AuthorizedParties } from "../../auth/authorized-parties.enum";
-import {
-  AllowAuthorizedParty,
-  AllowInactiveUser,
-  Groups,
-} from "../../auth/decorators";
+import { AllowAuthorizedParty, AllowInactiveUser } from "../../auth/decorators";
 import { ApiTags } from "@nestjs/swagger";
 import { BCeIDAccountTypeCodes } from "../../services/bceid/bceid.models";
+import { UserControllerService } from "..";
 
 @AllowAuthorizedParty(AuthorizedParties.institution)
 @Controller("users")
@@ -30,6 +26,7 @@ export class UserInstitutionsController extends BaseController {
   constructor(
     private readonly service: UserService,
     private readonly bceidService: BCeIDService,
+    private readonly userControllerService: UserControllerService,
   ) {
     super();
   }
@@ -67,42 +64,7 @@ export class UserInstitutionsController extends BaseController {
   async getAllBCeIDs(
     @UserToken() userToken: IUserToken,
   ): Promise<BCeIDAccountsAPIOutDTO> {
-    // Only business BCeID will execute a search on BCeID Web Services.
-    // Basic BCeID need to have the full user name provided to execute the
-    // getAccountDetail method on BCeID Web Services.
-    const account = await this.bceidService.getAccountDetails(
-      userToken.idp_user_name,
-      BCeIDAccountTypeCodes.Business,
-    );
-
-    if (!account) {
-      throw new UnprocessableEntityException(
-        "Not able to retrieve BCeID business account details for the current authenticated user.",
-      );
-    }
-
-    const searchOptions = new SearchAccountOptions();
-    searchOptions.requesterUserGuid = account.user.guid;
-    searchOptions.businessGuid = account.institution.guid;
-    const searchResult = await this.bceidService.searchBCeIDAccounts(
-      searchOptions,
-    );
-
-    const accounts = searchResult.accounts.map((account) => {
-      return {
-        guid: account.guid,
-        displayName: account.displayName,
-        email: account.email,
-        firstname: account.firstname,
-        surname: account.surname,
-        telephone: account.telephone,
-        userId: account.userId,
-      };
-    });
-
-    return {
-      accounts,
-    };
+    return await this.userControllerService.getAllBCeIDs(userToken);
   }
 
   @Get("/institution")
